organized baton.go a bit
