+++ baton/baton.go
@@ -1,9 +1,6 @@
 package baton
 
 import (
-	"encoding/json"
-	"fmt"
-	"log"
 	"net/http"
 	"time"
 
@@ -32,13 +29,6 @@ var upgrader = websocket.Upgrader{
 	},
 }
 
-var h = hub{
-	conns:      make(map[conn]struct{}),
-	register:   make(chan conn),
-	unregister: make(chan conn),
-	receive:    make(chan rawMsg),
-}
-
 // conn is the middleman between the websocket connection and the hub
 type conn struct {
 	ws   *websocket.Conn
@@ -46,7 +36,7 @@ type conn struct {
 }
 
 // readPump pumps messages from the websocket connection to the hub
-func (c conn) readPump() {
+func (c conn) readPump(h hub) {
 	defer func() {
 		h.unregister <- c
 		c.ws.Close()
@@ -94,84 +84,6 @@ func (c conn) writePump() {
 		case <-ticker.C:
 			if err := c.write(websocket.PingMessage, []byte{}); err != nil {
 				return
-			}
-		}
-	}
-}
-
-type command struct {
-	Module string      `json:"module"`
-	Call   string      `json:"call"`
-	Body   interface{} `json:"body"`
-}
-
-type rawMsg struct {
-	conn conn
-	data []byte
-}
-
-type msg struct {
-	conn conn
-	cmd  command
-}
-
-type hub struct {
-	conns      map[conn]struct{}
-	register   chan conn
-	unregister chan conn
-	receive    chan rawMsg
-}
-
-func (h hub) run() {
-	for {
-		select {
-		case c := <-h.register:
-			h.conns[c] = struct{}{}
-		case c := <-h.unregister:
-			if _, ok := h.conns[c]; ok {
-				delete(h.conns, c)
-				close(c.send)
-			}
-		case rawMsg := <-h.receive:
-			fmt.Println(rawMsg)
-			var cmd command
-			if err := json.Unmarshal(rawMsg.data, &cmd); err != nil {
-				fmt.Println("nu")
-				log.Println(err)
-				break
-			}
-			
-			msg := msg{rawMsg.conn, cmd}
-
-			fmt.Println(cmd)
-			resp := make(chan interface{})
-			module, ok := modules[msg.cmd.Module]
-			if !ok {
-				fmt.Println("the fuck is this module?")
-				break
-			}
-
-			go func() {
-				for {
-					r, ok := <-resp
-					if !ok {
-						close(resp)
-					}
-					bytes, err := json.Marshal(command{cmd.Module,cmd.Call,r}) //add in Module and Call info for client
-					if err != nil {
-					  log.Println("Error Marshaling")
-						log.Println(err)
-						break
-					}
-        
-					msg.conn.send <- bytes
-				}
-			}()
-
-			if err := module.RunCommand(cmd.Call, cmd.Body, resp); err != nil {
-			  fmt.Println(err)
-				fmt.Println("the fuck happened to this command?")
-				break
 			}
 		}
 	}
