+++ baton/handler.go
@@ -14,6 +14,13 @@ func Handler() *mux.Router {
 	return m
 }
 
+var h = hub{
+	conns:      make(map[conn]struct{}),
+	register:   make(chan conn),
+	unregister: make(chan conn),
+	receive:    make(chan rawMsg),
+}
+
 func Run() {
 	h.run()
 }
@@ -31,5 +38,5 @@ func serveWs(w http.ResponseWriter, r *http.Request) {
 	c := conn{send: make(chan []byte, 256), ws: ws}
 	h.register <- c
 	go c.writePump()
-	c.readPump()
+	c.readPump(h)
 }
